Create Okechukwu Favour kosisochukwu 2024714101
This python program implements a grading system using 'if' and 'elif' statement. It takes a student score as input and print the corresponding grade based on the following:
.70_100 Grade A
.60_69 Grade B
.50_59 Grade C
.45_49 Grade D
.40_44 Grade E
.0_39 Grade F
For each value score, the program also prints "YES" to confirm the grading was  successful. If the score is outside the valid range it prints "Invalid score" and "No".
Created by Okechukwu Favour kosisochukwu.
Reg no: 2024714101